refactor(4.2.6): fix to base global regexp
