+++ src/4.2.6.js
@@ -7,6 +7,11 @@
 例外は、住所や電話番号の区切りに使う場合です。
  */
 import {isUserWrittenNode} from "./util/node-util";
+import {matchCaptureGroupAll} from "./util/match-index";
+import regx from 'regx';
+import {japaneseRegExp} from "./util/regexp";
+import mergeMatches from "./util/merge-matches";
+const rx = regx("g");
 export default function (context) {
     let {Syntax, RuleError, report, getSource} = context;
     return {
@@ -15,18 +20,19 @@ export default function (context) {
                 return;
             }
             let text = getSource(node);
-            // 基本的にハイフン(-)を使用しません
-            var index = text.indexOf("-");
-            if (index !== -1) {
-                // 例外として、住所や電話番号の区切りにはハイフン(-)を使用できる
-                // ?-? を取り出して \d-\d ならOK、そうでないならダメ
-                var bufOfIndex = text.slice(index - 1, index + 2);
-                // [^A-Za-z0-9_] = \w
-                if (!/^\w\-\w$/.test(bufOfIndex)) {
-                    report(node, new RuleError(`原則として和文ではハイフン(-)を使用しません。
-例外は、住所や電話番号の区切りに使う場合です。`, index));
-                }
-            }
+            // 和文ではハイフン(-)を使用しません
+            // right
+            const rightMatches = matchCaptureGroupAll(text, rx`${japaneseRegExp}(\-)`);
+            // left
+            const leftMatches = matchCaptureGroupAll(text, rx`(\-)${japaneseRegExp}`);
+            const matches = mergeMatches(leftMatches, rightMatches);
+            matches.forEach(match => {
+                const {index} = match;
+                report(node, new RuleError(`原則として和文ではハイフン(-)を使用しません。
+例外は、住所や電話番号の区切りに使う場合です。`, {
+                    column: index
+                }))
+            });
         }
     };
 }
