+++ src/util/merge-matches.js
@@ -1,0 +1,17 @@
+// LICENSE : MIT
+"use strict";
+export default function mergeMatches(...aMatches) {
+    const results = [];
+    aMatches.forEach(matches => {
+        matches.forEach(targetMatch => {
+            const alreadyHave = results.some(match => {
+                const {text, index} = match;
+                return targetMatch.index === index && targetMatch.text === text;
+            });
+            if (!alreadyHave) {
+                results.push(targetMatch);
+            }
+        });
+    });
+    return results;
+}
