+++ test/4.2.6-test.js
@@ -21,11 +21,11 @@ tester.run("4.2.6.ハイフン(-)", rule, {
             ]
         },
         {
-            text: "090 - 1234 - 5678",
+            text: "A市-B市",
             errors: [
                 {
                     message: "原則として和文ではハイフン(-)を使用しません。\n例外は、住所や電話番号の区切りに使う場合です。",
-                    column: 5
+                    column: 3
                 }
             ]
         }
